% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutsneedle.R
\name{mutsneedle}
\alias{mutsneedle}
\title{Mutsneedle}
\usage{
mutsneedle(title = " Mutation Needle Plot", domains = NULL,
  mutdata = NULL, colorMap = NULL, gene = "TP53",
  transcript = "ENST032423423", maxlength = 800, width = NULL,
  height = NULL, elementId = NULL)
}
\description{
Interactive mutation lolipop chart from the Muts-Needle-Plot package from R
}
\examples{
Example - with data


#Get some mutation and region data
data <- exampleMutationData()

regiondata <- exampleRegionData()

#Make the lolipop chart
mutsneedle(mutdata=data,domains=regiondata)

# Example 2 - in a Shiny app


library(shiny)
shinyApp(
  ui = mutsneedleOutput("id",width=800,height=500),
  server = function(input, output) {
    output$id <- renderMutsneedle(
       data <- exampleMutationData()
       regiondata <- exampleRegionData()
      mutsneedle(mutdata=data,domains=regiondata)
    )
  }
)

}
